# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Levannel API",
    "version": "1.0.0",
    "description": "API REST de gestion d’agence de voyage développée avec Node.js, Express, TypeScript et PostgreSQL."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    },
    {
      "url": "https://levanel-api.onrender.com",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/services": {
      "post": {
        "summary": "Créer un service",
        "description": "Crée un nouveau service métier. Une référence interne est automatiquement générée.",
        "tags": [
          "Services"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Billetterie"
                  },
                  "description": {
                    "type": "string",
                    "example": "Gestion des billets"
                  },
                  "color": {
                    "type": "string",
                    "example": "#2563EB"
                  },
                  "icon": {
                    "type": "string",
                    "example": "plane"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service créé avec succès"
          },
          "400": {
            "description": "Données invalides"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès réservé aux administrateurs, managers et agents"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "get": {
        "summary": "Liste des services",
        "description": "Retourne tous les services disponibles.",
        "tags": [
          "Services"
        ],
        "responses": {
          "200": {
            "description": "Liste récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Services retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "example": "Billetterie"
                          },
                          "initial": {
                            "type": "string",
                            "example": "BIL"
                          },
                          "description": {
                            "type": "string",
                            "example": "Gestion des billets"
                          },
                          "color": {
                            "type": "string",
                            "example": "#2563EB"
                          },
                          "icon": {
                            "type": "string",
                            "example": "plane"
                          },
                          "is_active": {
                            "type": "boolean",
                            "example": true
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-20T12:50:38.156Z"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2026-05-20T12:50:38.156Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/services/{id}": {
      "put": {
        "summary": "Modifier un service",
        "description": "Met à jour les informations d'un service.",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Identifiant du service"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Billetterie"
                  },
                  "description": {
                    "type": "string",
                    "example": "Gestion des billets internationaux"
                  },
                  "color": {
                    "type": "string",
                    "example": "#2563EB"
                  },
                  "icon": {
                    "type": "string",
                    "example": "plane"
                  },
                  "is_active": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service modifié avec succès"
          },
          "400": {
            "description": "Données invalides"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès réservé aux administrateurs, managers et agents"
          },
          "404": {
            "description": "Service introuvable"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un service",
        "description": "Effectue une suppression logique d'un service.",
        "tags": [
          "Services"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifiant du service"
          }
        ],
        "responses": {
          "200": {
            "description": "Service supprimé avec succès"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès réservé aux administrateurs et managers"
          },
          "404": {
            "description": "Service introuvable"
          }
        }
      }
    },
    "/api/service-requests": {
      "post": {
        "summary": "Créer une demande de service",
        "description": "Crée une nouvelle demande de service associée à un client, un service et éventuellement un partenaire.",
        "tags": [
          "Service Requests"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "client_id",
                  "service_id",
                  "request_type",
                  "currency"
                ],
                "properties": {
                  "client_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "CLIENT_UUID"
                  },
                  "service_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "SERVICE_UUID"
                  },
                  "partner_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "PARTNER_UUID"
                  },
                  "contract_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "CONTRACT_UUID"
                  },
                  "request_type": {
                    "type": "string",
                    "example": "ticketing"
                  },
                  "currency": {
                    "type": "string",
                    "example": "USD"
                  },
                  "observation": {
                    "type": "string",
                    "example": "Mission Bruxelles juin 2026"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demande créée avec succès"
          },
          "400": {
            "description": "Données invalides"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès refusé"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "get": {
        "summary": "Liste des demandes de services",
        "description": "Retourne la liste paginée des demandes de services.",
        "tags": [
          "Service Requests"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            },
            "description": "Numéro de page"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            },
            "description": "Nombre d'éléments par page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string",
              "example": "Bruxelles"
            },
            "description": "Recherche par référence, client ou observation"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "example": "pending"
            },
            "description": "Filtrer par statut"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Requests retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "request_reference": {
                                "type": "string",
                                "example": "SR-1779449061291-1711"
                              },
                              "client_id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "client_name": {
                                "type": "string",
                                "example": "Equity BCDC"
                              },
                              "service_id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "service_name": {
                                "type": "string",
                                "example": "Billetterie"
                              },
                              "partner_id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "partner_name": {
                                "type": "string",
                                "example": "Congo Airways"
                              },
                              "contract_id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "request_type": {
                                "type": "string",
                                "example": "ticketing"
                              },
                              "status": {
                                "type": "string",
                                "example": "pending"
                              },
                              "total_amount": {
                                "type": "string",
                                "example": "1900.00"
                              },
                              "amount_paid": {
                                "type": "string",
                                "example": "500.00"
                              },
                              "remaining_amount": {
                                "type": "number",
                                "example": 1400
                              },
                              "service_revenue": {
                                "type": "string",
                                "example": "51.40"
                              },
                              "external_cost": {
                                "type": "string",
                                "example": "1848.60"
                              },
                              "currency": {
                                "type": "string",
                                "example": "USD"
                              },
                              "observation": {
                                "type": "string",
                                "example": "Mission Bruxelles juin 2026"
                              },
                              "requested_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-05-22T11:24:21.331Z"
                              },
                              "completed_at": {
                                "type": "string",
                                "nullable": true,
                                "example": null
                              },
                              "created_at": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "updated_at": {
                                "type": "string",
                                "format": "date-time"
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer",
                          "example": 1
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "limit": {
                          "type": "integer",
                          "example": 10
                        },
                        "total_pages": {
                          "type": "integer",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
    "/api/service-requests/{id}": {
      "put": {
        "summary": "Modifier une demande de service",
        "description": "Met à jour une demande de service existante.",
        "tags": [
          "Service Requests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Identifiant de la demande"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "completed"
                  },
                  "observation": {
                    "type": "string",
                    "example": "Voyage confirmé et billet émis"
                  },
                  "completed_at": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Demande modifiée avec succès"
          },
          "400": {
            "description": "Données invalides"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès réservé aux administrateurs et managers"
          },
          "404": {
            "description": "Demande introuvable"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      },
      "delete": {
        "summary": "Supprimer une demande de service",
        "description": "Effectue une suppression logique d'une demande de service.",
        "tags": [
          "Service Requests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Identifiant de la demande"
          }
        ],
        "responses": {
          "200": {
            "description": "Demande supprimée avec succès"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Accès réservé aux administrateurs et managers"
          },
          "404": {
            "description": "Demande introuvable"
          },
          "500": {
            "description": "Erreur serveur"
          }
        }
      }
    },
...
  },
...
}